+++ hash-sha4.h
@@ -43,7 +43,7 @@ void sha4_starts(sha4_ctx * nonnull ctx, int is384) __leaf;
  * \param ilen     length of the input data
  */
 void sha4_update(sha4_ctx * nonnull ctx, const void * nonnull input,
-                 int ilen) __leaf;
+                 ssize_t ilen) __leaf;
 
 /**
  * \brief          SHA-512 final digest
@@ -61,7 +61,7 @@ void sha4_finish(sha4_ctx * nonnull ctx, byte output[64]) __leaf;
  * \param output   SHA-384/512 checksum result
  * \param is384    0 = use SHA512, 1 = use SHA384
  */
-void sha4(const void * nonnull input, int ilen, byte output[64], int is384)
+void sha4(const void * nonnull input, ssize_t ilen, byte output[64], int is384)
     __leaf;
 
 /**
@@ -95,7 +95,7 @@ void sha4_hmac_starts(sha4_ctx * nonnull ctx, const void * nonnull key,
  * \param ilen     length of the input data
  */
 void sha4_hmac_update(sha4_ctx * nonnull ctx, const void * nonnull input,
-                      int ilen) __leaf;
+                      ssize_t ilen) __leaf;
 
 /**
  * \brief          SHA-512 HMAC final digest
@@ -116,7 +116,7 @@ void sha4_hmac_finish(sha4_ctx * nonnull ctx, byte output[64]) __leaf;
  * \param is384    0 = use SHA512, 1 = use SHA384
  */
 void sha4_hmac(const void * nonnull key, int keylen,
-               const void * nonnull input, int ilen,
+               const void * nonnull input, ssize_t ilen,
                byte output[64], int is384) __leaf;
 
 #ifdef __cplusplus
